+++ log/kafka/kafka.go
@@ -14,16 +14,20 @@ import (
 import (
 	// "github.com/Shopify/sarama"
 	"github.com/Shopify/sarama"
+	sc "github.com/bsm/sarama-cluster"
 	"github.com/wvanbergen/kazoo-go"
 )
 
-// type ProducerError struct {
-// 	Msg *ProducerMessage
-// 	Err error
-// }
 type (
 	// Consumer will invoke @ProduceMessageCallback when got message
-	ConsumerMessageCallback func(*sarama.ConsumerMessage) error
+	// @msg: consumer message
+	// @preOffset: @msg's previous message's offset in the same partition.
+	//             If @msg is this partition's first message, its preOffset is 0.
+	ConsumerMessageCallback func(msg *sarama.ConsumerMessage, preOffset int64)
+	// Consumer will invoke @ConsumerErrorCallback when got error
+	ConsumerErrorCallback func(error)
+	// Consumer will invoke @ConsumerNotification when got notification
+	ConsumerNotificationCallback func(*sc.Notification)
 	// AsyncProducer will invoke @ProduceMessageCallback when got sucess message response.
 	ProducerMessageCallback func(*sarama.ProducerMessage)
 	// AsyncProducer will invoke @ProduceErrorCallback when got error message response